stores: update latest_host in DeleteHostSector even if no links to delete were found

+++ stores/metadata.go
@@ -2,7 +2,6 @@ package stores
 
 import (
 	"context"
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"math"
@@ -1324,16 +1323,16 @@ func (s *SQLStore) CopyObject(ctx context.Context, srcBucket, dstBucket, srcPath
 
 func (s *SQLStore) DeleteHostSector(ctx context.Context, hk types.PublicKey, root types.Hash256) error {
 	return s.retryTransaction(func(tx *gorm.DB) error {
-		// Fetch sectors to delete.
+		// Fetch contract_sectors to delete.
 		var sectors []dbContractSector
-		err := tx.Raw(fmt.Sprintf(`
+		err := tx.Raw(`
 			SELECT contract_sectors.*
 			FROM contract_sectors
 			INNER JOIN sectors s ON s.id = contract_sectors.db_sector_id
 			INNER JOIN contracts c ON c.id = contract_sectors.db_contract_id
 			INNER JOIN hosts h ON h.id = c.host_id
-			WHERE s.root = x'%s' AND h.public_key = x'%s'
-			`, hex.EncodeToString(root[:]), hex.EncodeToString(hk[:]))).
+			WHERE s.root = ? AND h.public_key = ?
+			`, root[:], publicKey(hk)).
 			Scan(&sectors).
 			Error
 		if err != nil {
@@ -1342,45 +1341,49 @@ func (s *SQLStore) DeleteHostSector(ctx context.Context, hk types.PublicKey, roo
 			return fmt.Errorf("root %v for host %v not found: %w", root, hk, gorm.ErrRecordNotFound)
 		}
 
-		// Update the affected slabs.
-		var sectorIDs []uint
-		uniqueIDs := make(map[uint]struct{})
-		for _, s := range sectors {
-			if _, exists := uniqueIDs[s.DBSectorID]; !exists {
-				uniqueIDs[s.DBSectorID] = struct{}{}
-				sectorIDs = append(sectorIDs, s.DBSectorID)
+		if len(sectors) > 0 {
+			// Update the affected slabs.
+			var sectorIDs []uint
+			uniqueIDs := make(map[uint]struct{})
+			for _, s := range sectors {
+				if _, exists := uniqueIDs[s.DBSectorID]; !exists {
+					uniqueIDs[s.DBSectorID] = struct{}{}
+					sectorIDs = append(sectorIDs, s.DBSectorID)
+				}
+			}
+			err = tx.Exec("UPDATE slabs SET health_valid = 0 WHERE id IN (SELECT db_slab_id FROM sectors WHERE id IN (?))", sectorIDs).Error
+			if err != nil {
+				return fmt.Errorf("failed to invalidate slab health: %w", err)
+			}
+
+			// Delete contract_sectors.
+			res := tx.Delete(&sectors)
+			if err := res.Error; err != nil {
+				return fmt.Errorf("failed to delete contract sectors: %w", err)
+			} else if res.RowsAffected != int64(len(sectors)) {
+				return fmt.Errorf("expected %v affected rows but got %v", len(sectors), res.RowsAffected)
 			}
 		}
-		err = tx.Exec("UPDATE slabs SET health_valid = 0 WHERE id IN (SELECT db_slab_id FROM sectors WHERE id IN (?))", sectorIDs).Error
-		if err != nil {
-			return fmt.Errorf("failed to invalidate slab health: %w", err)
-		}
 
-		// Delete contract_sectors.
-		res := tx.Delete(&sectors)
-		if err := res.Error; err != nil {
-			return fmt.Errorf("failed to delete contract sectors: %w", err)
-		} else if res.RowsAffected != int64(len(sectors)) {
-			return fmt.Errorf("expected %v affected rows but got %v", len(sectors), res.RowsAffected)
-		}
-
-		// Fetch the sectors and update the latest_host field.
-		var dbSectors []dbSector
-		err = tx.Where("id IN (?)", sectorIDs).
+		// Fetch the sector and update the latest_host field if the host for
+		// which we remove the sector is the latest_host.
+		var sector dbSector
+		err = tx.Where("root", root[:]).
 			Preload("Contracts.Host").
-			Find(&dbSectors).
+			Find(&sector).
 			Error
 		if err != nil {
 			return fmt.Errorf("failed to fetch sectors: %w", err)
 		}
-		for i, sector := range dbSectors {
+		if sector.LatestHost == publicKey(hk) {
 			if len(sector.Contracts) == 0 {
-				dbSectors[i].LatestHost = publicKey{} // no more hosts
+				sector.LatestHost = publicKey{} // no more hosts
 			} else {
-				dbSectors[i].LatestHost = sector.Contracts[len(sector.Contracts)-1].Host.PublicKey // most recent contract
+				sector.LatestHost = sector.Contracts[len(sector.Contracts)-1].Host.PublicKey // most recent contract
 			}
+			return tx.Save(sector).Error
 		}
-		return tx.Save(dbSectors).Error
+		return nil
 	})
 }
